Enable processing from the registered schema in the generator package's renderer (for internal use only)
