+++ packages/generator/src/renderUtils.ts
@@ -1,0 +1,42 @@
+import { PDFPage, degrees, } from '@pdfme/pdf-lib';
+import { Schema, TextSchema, Alignment, mm2pt } from '@pdfme/common';
+import { hex2RgbColor } from "./colorUtils"
+
+export const calcX = (x: number, alignment: Alignment, boxWidth: number, textWidth: number) => {
+  let addition = 0;
+  if (alignment === 'center') {
+    addition = (boxWidth - textWidth) / 2;
+  } else if (alignment === 'right') {
+    addition = boxWidth - textWidth;
+  }
+
+  return mm2pt(x) + addition;
+};
+
+export const calcY = (y: number, height: number, itemHeight: number) => height - mm2pt(y) - itemHeight;
+
+export const renderBackgroundColor = (arg: {
+  templateSchema: TextSchema;
+  page: PDFPage;
+  pageHeight: number;
+}) => {
+  const { templateSchema, page, pageHeight } = arg;
+  if (!templateSchema.backgroundColor) return;
+  const { width, height } = convertSchemaDimensionsToPt(templateSchema);
+  const color = hex2RgbColor(templateSchema.backgroundColor);
+  page.drawRectangle({
+    x: calcX(templateSchema.position.x, 'left', width, width),
+    y: calcY(templateSchema.position.y, pageHeight, height),
+    width,
+    height,
+    color,
+  });
+};
+
+export const convertSchemaDimensionsToPt = (schema: Schema) => {
+  const width = mm2pt(schema.width);
+  const height = mm2pt(schema.height);
+  const rotate = degrees(schema.rotate ? schema.rotate : 0);
+
+  return { width, height, rotate };
+};
